Fixed convertToRational() function, now all variables are double as long as it is possible, and the precision of the last rational number is tuned to the maximum allowed by the exif library.
git-svn-id: file:///home/pete/devel/svn/easygeotagger/trunk@57 622b0c1a-4c31-43e3-84e9-cb026a211916

+++ easygeotagger/src/core/egtexifio.h
@@ -37,8 +37,8 @@ public:
   EgtExifIO( QString );
   EgtExifIO( const QModelIndex& );
   
-  float latitude();
-  float longitude();
+  double latitude();
+  double longitude();
   
   bool hasGpsExif();
   bool isValidImage();
